Move test-gen-lib to it's own submodule
