+++ pom.xml
@@ -154,6 +154,11 @@
       </dependency>
       <dependency>
         <groupId>com.dylibso.chicory</groupId>
+        <artifactId>test-gen-lib</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>com.dylibso.chicory</groupId>
         <artifactId>wabt</artifactId>
         <version>${project.version}</version>
       </dependency>
@@ -838,6 +843,7 @@
         <module>machine-tests</module>
         <module>runtime</module>
         <module>runtime-tests</module>
+        <module>test-gen-lib</module>
         <module>test-gen-plugin</module>
         <module>wabt</module>
         <module>wasi</module>
