+++ test-gen-lib/src/main/java/com/dylibso/chicory/testgen/TestSuiteDownloader.java
@@ -1,4 +1,4 @@
-package com.dylibso.chicory.maven;
+package com.dylibso.chicory.testgen;
 
 import java.io.File;
 import java.io.IOException;
@@ -11,31 +11,23 @@ import java.nio.file.Paths;
 import java.util.Comparator;
 import java.util.stream.Stream;
 import net.lingala.zip4j.ZipFile;
-import org.apache.maven.plugin.logging.Log;
 
 public class TestSuiteDownloader {
 
-    private final Log log;
-
-    TestSuiteDownloader(Log log) {
-        this.log = log;
-    }
+    TestSuiteDownloader() {}
 
     public void downloadTestsuite(
             String testSuiteRepo, String testSuiteRepoRef, File testSuiteFolder)
             throws IOException {
         if (testSuiteFolder.exists()
                 && testSuiteFolder.list((dir, name) -> name.endsWith(".wast")).length == 0) {
-            log.warn("Testsuite folder exists but looks corrupted, replacing.");
             try (Stream<Path> files =
                     Files.walk(testSuiteFolder.toPath()).sorted(Comparator.reverseOrder())) {
                 files.map(Path::toFile).forEach(File::delete);
             }
-        } else {
-            log.debug("Testsuite detected, using the cached version.");
         }
+
         if (!testSuiteFolder.exists()) {
-            log.warn("Cloning the testsuite at ref: " + testSuiteRepoRef);
             // target URL:
             // https://github.com/WebAssembly/testsuite/archive/c2a67a575ddc815ff2212f68301d333e5e30a923.tar.gz
             // https://docs.github.com/en/repositories/working-with-files/using-files/downloading-source-code-archives#source-code-archive-urls
@@ -54,7 +46,6 @@ public class TestSuiteDownloader {
             } finally {
                 con.disconnect();
             }
-            log.warn("Cloned the testsuite at ref: " + testSuiteRepoRef);
         }
     }
 }
