+++ test-gen-lib/src/main/java/com/dylibso/chicory/testgen/wast/LaneType.java
@@ -1,9 +1,13 @@
-package com.dylibso.chicory.maven.wast;
+package com.dylibso.chicory.testgen.wast;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonValue;
 
-public enum WasmValueType {
+public enum LaneType {
+    @JsonProperty("i8")
+    I8("i8"),
+    @JsonProperty("i16")
+    I16("i16"),
     @JsonProperty("i32")
     I32("i32"),
     @JsonProperty("i64")
@@ -11,17 +15,11 @@ public enum WasmValueType {
     @JsonProperty("f32")
     F32("f32"),
     @JsonProperty("f64")
-    F64("f64"),
-    @JsonProperty("v128")
-    V128("v128"),
-    @JsonProperty("externref")
-    EXTERN_REF("externref"),
-    @JsonProperty("funcref")
-    FUNC_REF("funcref");
+    F64("f64");
 
     private final String value;
 
-    WasmValueType(String value) {
+    LaneType(String value) {
         this.value = value;
     }
